migrate: move promoBanner to components directory
- Migrated promoBanner component files to src/components/promoBanner/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Updated import statements:
  - src/templates/common/pageLayout/index.tsx from @/templates/common/banners/promoBanner to @/components/promoBanner/template
  - src/templates/common/pageLayout/index.test.tsx from @/templates/common/banners/promoBanner to @/components/promoBanner/template
  - template.test.tsx: from external import to ./template
- Removed empty src/templates/common/banners/promoBanner/ directory
- All TypeScript compilation checks passing
- All tests passing (2 tests for PromoBanner component)

This promoBanner component renders promotional banners using VA Component Library.
It uses shared types from the banner component's formatted-type for PromoBanner interface.
