+++ src/components/promoBanner/template.tsx
