+++ src/templates/common/pageLayout/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from 'test-utils'
 import '@testing-library/jest-dom'
 import { PageLayout, formatBannerType } from './index'
-import { PromoBanner } from '@/templates/common/banners/promoBanner'
+import { PromoBanner } from '@/components/promoBanner/template'
 import { Banner } from '@/components/banner/template'
 import { BANNER_RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
